ARM: dts: exynos: override pins by label in Peach Pit
Using node paths to extend or override a device tree node is error
prone.  If there was a typo error, a new node will be created instead of
extending the existing node.  This will lead to run-time errors that
could be hard to detect.

A mistyped label on the other hand, will cause a dtc compile error
(during build time).

Signed-off-by: Krzysztof Kozlowski <krzysztof.kozlowski@canonical.com>
Reviewed-by: Alim Akhtar <alim.akhtar@samsung.com>
Link: https://lore.kernel.org/r/20220111201426.326777-6-krzysztof.kozlowski@canonical.com

[ upstream commit: 213e19b4f6752753715bb8e8fa46b8f0cc1dd68d ]