arm64: dts: renesas: ulcb-kf: Add 9-asix sensor device
This adds nodes for the lsm9ds0 sensor installed on the KF board.

With this patch, the sensor data becomes available over the IIO sysfs
interface.

The interrupt definition is not added yet, because the interrupt lines
of lsm9ds0 are pulled to VCC on the board, which implies a need for
active-low configuration. But the st_sensors drivers currently can't
work with active-low interrupts on this chip.

Signed-off-by: Nikita Yushchenko <nikita.yoush@cogentembedded.com>
Link: https://lore.kernel.org/r/20220112205205.4082026-1-nikita.yoush@cogentembedded.com
Signed-off-by: Geert Uytterhoeven <geert+renesas@glider.be>

[ upstream commit: c705c871106e35221f486765fcc3d978a4434c38 ]