ARM: dts: ux500: Update AB850[05] nodes
The new YAML device tree bindings gives new and proper names
to several of the AB850[05] nodes and redefines the way we
use numbering on PWMs to use reg.

Update all the DTS nodes accordingly.

Add the missing thermal node to the AB8505.

Drop the debugfs nodes because these are not real devices.

Signed-off-by: Linus Walleij <linus.walleij@linaro.org>

[ upstream commit: 218b2f88853fd448a6229b185e222e565716d860 ]