ARM: dts: NSP: MX6X: correct LED function types
Currently, the amber LED will remain always on. This is due to a
misinterpretation of the LED sub-node properties, where-by "default-state"
was used to indicate the initial state when powering on the device. When in
use, however, this resulted in the amber LED always being on. Instead change
this to only indicate a fault state.

Assign LED_FUNCTION_POWER to the green PWM LED.

These changes bring the MX64/65 in line with the MR32's devicetree.

Signed-off-by: Matthew Hagan <mnhagan88@gmail.com>
Signed-off-by: Florian Fainelli <f.fainelli@gmail.com>

[ upstream commit: 482c85c7fc95c572d368b2214b9e9d2c4a2e5789 ]